Added support for asynchronous proxies (early snapshot).
[This is an early snapshot to make these changes available for review.]

Modified the proxy code generator to optionally create code
that makes non-blocking method invocations and invokes callbacks
when replies are asynchronously received. Also completed/fixed
the PendingCall class, and the support in the library for async
invocation and reply handling.

For a description of the approach taken see the following document:
https://docs.google.com/a/google.com/document/d/1m2h8klOYSna7-TkNg9-_fT3_GhKIgB3UJVDDxhjwojU/edit?hl=en_US

BUG=chromium-os:17583
TEST=Added a new sample server and a client that uses asynchronous
calls.

Change-Id: I8b4bfb628f2a0515a91c5821d20e878fb76dc7e7
Reviewed-on: https://gerrit.chromium.org/gerrit/10244
Reviewed-by: Darin Petkov <petkov@chromium.org>
Tested-by: Eric Shienbrood <ers@chromium.org>