Use ctemplate for proxy and adaptor code generation.
Changed the stub generator to expand a template using
the ctemplate library, instead of embedding all the code
to be generated in the generator itself. Provides a
separation between "presentation" and logic.

generate_adaptor.cpp is no longer needed, as the same code
can be used with different templates to generate both proxy
code and adaptor code. generate_proxy.cpp was renamed to
generate_stubs.cpp.

BUG=chromium-os:24140
TEST=I built and ran on my workstation, outside of the chromeos
chroot. I've also emerge-d it inside the chroot and used it
to build and successfully run shill.

Change-Id: Id5a6b91f646935ed80be9bd50d791904b38b3b57
Reviewed-on: https://gerrit.chromium.org/gerrit/15665
Tested-by: Eric Shienbrood <ers@chromium.org>
Commit-Ready: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Tested-by: Jason Glasgow <jglasgow@chromium.org>