Fixed hang due to infinite loop in dbus-c++ async method reply handlers.
If an async method reply handler makes a non-blocking dbus call, it
is possible that the async reply to the second call could be handled
before handling of the original reply is completed. In that case, when
the stub code for the original reply callback tries to remove the
PendingCall it was handling, the removal code enters an infinite loop.
This fixes that.

BUG=chromium-os:25603
TEST=Discovered while working on making more client operations
asynchronous in shill. The problem is not consistently repeatable,
but will occur once every few times the cellular device is enabled.
The observed behavior when this bug is present is that shill hangs
and becomes unresponsive.

Change-Id: I94ff5156b71fd6e140fb3b8214c84c5be1e390bc
Reviewed-on: https://gerrit.chromium.org/gerrit/14982
Tested-by: Eric Shienbrood <ers@chromium.org>
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Commit-Ready: Eric Shienbrood <ers@chromium.org>