dbus-c++: Ensure correct size of opaque MessageIter::_iter member.
The MessageIter::_iter data member is supposed to match size of DBusMessageIter,
however, its size was hardcoded in dbus-c++/message.h to avoid including
dbus/dbus.h. In the current build this resulted in sizeof(_iter)=68, less than
sizeof(DBusMessageIter)=72.

This CL dynamically allocates a private structure that includes a
DBusMessageIter member thus ensuring that the member size stays in
sync with the struct. The patch also adds an explicit copy constructor
and assignment operator so that MessageIter objects are copied
properly.

A couple of additional changes:

- Reset MessageIter data members to NULL in the destructor.
- Avoid implicit conversion between bool and dbus_bool_t.

BUG=chromium-os:30096
TEST=shill unit tests, build and run a test image

Change-Id: I9b56449f94928e41395ed44228f94eaf203e440a
Reviewed-on: https://gerrit.chromium.org/gerrit/21749
Tested-by: Darin Petkov <petkov@chromium.org>
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Darin Petkov <petkov@chromium.org>